Added form to critique page, fixed header css
The header no longer looks super buggy anymore! The background for it wasn't reaching all the way because there were margins on it, so when you scrolled the content peeked behind it. also there was no height parameter on it, so the content would start from 0 instead of under the header.